Restructured directories, renamed files, updated imports, tweaked Header, & updated NavMenu (Navbar).
