+++ frontend/src/app/page.tsx
@@ -1,11 +1,10 @@
 import Image from "next/image";
-import ListingCard from "@/components/listings/ListingCard";
-import Navbar from "@/components/shared/Navbar";
-import Header from "@/components/header/header";
+import ListingCard from "@/components/ListingCard";
+import Header from "@/components/Header";
 export default function Home() {
   return (
-    <div>
-      <Header />
-    </div>
+    <>
+      <Header/>
+    </>
   );
 }
