+++ frontend/src/components/ListingCard.tsx
